refactor samples, and more tests
